+++ mobile/android/android-components/components/feature/downloads/src/main/java/mozilla/components/feature/downloads/AbstractFetchDownloadService.kt
@@ -70,6 +70,7 @@ import mozilla.components.feature.downloads.facts.emitNotificationPauseFact
 import mozilla.components.feature.downloads.facts.emitNotificationResumeFact
 import mozilla.components.feature.downloads.facts.emitNotificationTryAgainFact
 import mozilla.components.support.base.log.logger.Logger
+import mozilla.components.support.ktx.kotlin.ifNullOrEmpty
 import mozilla.components.support.ktx.kotlin.sanitizeURL
 import mozilla.components.support.ktx.kotlinx.coroutines.throttleLatest
 import mozilla.components.support.utils.DownloadUtils
@@ -461,7 +462,7 @@ abstract class AbstractFetchDownloadService : Service() {
                     title = fileName,
                     description = fileName,
                     isMediaScannerScannable = true,
-                    mimeType = download.contentType ?: "*/*",
+                    mimeType = getSafeContentType(context, download.filePath, download.contentType),
                     path = file.absolutePath,
                     length = download.contentLength ?: file.length(),
                     // Only show notifications if our channel is blocked
@@ -885,7 +886,9 @@ abstract class AbstractFetchDownloadService : Service() {
     internal fun useFileStreamScopedStorage(download: DownloadState, block: (OutputStream) -> Unit) {
         val values = ContentValues().apply {
             put(MediaStore.Downloads.DISPLAY_NAME, download.fileName)
-            put(MediaStore.Downloads.MIME_TYPE, download.contentType ?: "*/*")
+            put(MediaStore.Downloads.MIME_TYPE,
+                getSafeContentType(context, download.filePath, download.contentType)
+            )
             put(MediaStore.Downloads.SIZE, download.contentLength)
             put(MediaStore.Downloads.IS_PENDING, 1)
         }
@@ -968,12 +971,7 @@ abstract class AbstractFetchDownloadService : Service() {
         fun openFile(context: Context, filePath: String, contentType: String?): Boolean {
             // Create a new file with the location of the saved file to extract the correct path
             // `file` has the wrong path, so we must construct it based on the `fileName` and `dir.path`s
-            val fileLocation = File(filePath)
-            val constructedFilePath = FileProvider.getUriForFile(
-                context,
-                context.packageName + FILE_PROVIDER_EXTENSION,
-                fileLocation
-            )
+            val constructedFilePath = getFilePathUri(context, filePath)
 
             val newIntent = Intent(ACTION_VIEW).apply {
                 setDataAndType(constructedFilePath, getSafeContentType(context, constructedFilePath, contentType))
@@ -994,13 +992,23 @@ abstract class AbstractFetchDownloadService : Service() {
             val resultContentType = if (!contentTypeFromFile.isNullOrEmpty()) {
                 contentTypeFromFile
             } else {
-                if (!contentType.isNullOrEmpty()) {
-                    contentType
-                } else {
-                    "*/*"
-                }
+                contentType.ifNullOrEmpty { "*/*" }
             }
-            return (DownloadUtils.sanitizeMimeType(resultContentType) ?: "*/*")
+            return DownloadUtils.sanitizeMimeType(resultContentType).ifNullOrEmpty { "*/*" }
+        }
+
+        @VisibleForTesting
+        internal fun getSafeContentType(context: Context, filePath: String, contentType: String?): String {
+            return getSafeContentType(context, getFilePathUri(context, filePath), contentType)
+        }
+
+        @VisibleForTesting
+        internal fun getFilePathUri(context: Context, filePath: String): Uri {
+            return FileProvider.getUriForFile(
+                context,
+                context.packageName + FILE_PROVIDER_EXTENSION,
+                File(filePath)
+            )
         }
 
         private const val FILE_PROVIDER_EXTENSION = ".feature.downloads.fileprovider"